создание модуля user_interface.py и перенос программы
